feat(): modularize argument parsing into dedicated files
